requirements.txt
numpy #==1.16.0
opencv-python #==4.0.0.21
pdf2image #==1.4.0
Pillow #==8.1.1
pytesseract #==0.2.6
XlsxWriter #==1.1.2
django #==5.1.1
